Only output and object file if directly requested via `-c, `-r` or `-shared`
We have a `link_to_object` mode which can be triggered explicitly via
the `-r` flag, but is triggered implicitly for any unknown file
extension.  This is for compatibility with build system that link to
`.so` or `.dylib` files.

However, we can be more precise about this and perform `link_to_object`
if the user explicitly passes `-shared` or `-r`.

Without `-shared` we can/should just assume that the user wants to link
to a JS file.  This allows linking to files with arbitrary extensions.
For example I can do `emcc foo.c -o foo.txt` and I will actually get a
JS file called `foo.txt`.  Previously you would see: `Assuming object file
output in the absence of `-c`, based on output filename`

This removes some rather unhelpful errors and warnings that we were
needless generating when this `link_to_object` mode was automatically
triggered.

It also allowed me to do some refactoring to move js_target and
final_suffix declarations closer to their actual first use.

See: #11753
Replaces: #7697